moved header and logo to component
