+++ app/components/layout/index.ts
@@ -3,4 +3,5 @@ export * from './content'
 export * from './error-boundary'
 export * from './footer'
 export * from './form'
+export * from './header'
 export * from './side-nav'
