+++ app/components/templates/index.ts
@@ -2,6 +2,7 @@ export * from './checkbox-field'
 export * from './errors'
 export * from './field'
 export * from './hidden-submit'
+export * from './logo'
 export * from './progress-bar'
 export * from './search-bar'
 export * from './spacer'
